+++ src/examples/EX.ts
@@ -1,4 +1,4 @@
 
 console.log('----------------------------- EX.ts ------------------');
 
-require('./cancel');
+require('./restart');