+++ CHANGELOG.md
@@ -12,6 +12,12 @@ and here
 
 ### Changed
 -->
+## Release 2.1.5 -- 2024-03
+- Added engine.startEvent to API
+### Enhanced cli
+- support double quotes, so parameters like: `name "Buy Used Car" ` can now work
+- support backspace on input
+
 ## Release 2.1.0 -- 2024-03
 - Added engine.restart